src/main/java/com/siyanda/bankatmweb/repository/AccountRepository.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.siyanda.bankatmweb.repository;

import com.siyanda.bankatmweb.domain.Account;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.stereotype.Repository;

/**
 *
 * @author inathi-zenande
 */
@Repository
public interface AccountRepository extends JpaRepository<Account, Long>{

    public void save(Account account);

    public Account findOne(Long id);    

    public void delete(Account account);

    public void deleteAll();
}
